back/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
add-trailing-comma==3.0.0
    # via -r requirements.in
aiohttp==3.8.5
    # via -r requirements.in
aiosignal==1.3.1
    # via aiohttp
alembic==1.11.1
    # via -r requirements.in
annotated-types==0.5.0
    # via pydantic
anyio==3.7.1
    # via
    #   httpcore
    #   starlette
apscheduler==3.10.1
    # via -r requirements.in
async-timeout==4.0.2
    # via aiohttp
asyncpg==0.28.0
    # via -r requirements.in
attrs==23.1.0
    # via aiohttp
black==23.7.0
    # via -r requirements.in
certifi==2023.7.22
    # via
    #   httpcore
    #   httpx
charset-normalizer==3.2.0
    # via aiohttp
click==8.1.6
    # via
    #   black
    #   uvicorn
fastapi==0.100.1
    # via -r requirements.in
flake8==6.1.0
    # via -r requirements.in
frozenlist==1.4.0
    # via
    #   aiohttp
    #   aiosignal
greenlet==2.0.2
    # via sqlalchemy
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==0.17.3
    # via httpx
httpx==0.24.1
    # via -r requirements.in
idna==3.4
    # via
    #   anyio
    #   httpx
    #   yarl
isort==5.12.0
    # via -r requirements.in
mako==1.2.4
    # via alembic
markupsafe==2.1.3
    # via mako
mccabe==0.7.0
    # via flake8
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via black
packaging==23.1
    # via black
pathspec==0.11.2
    # via black
platformdirs==3.10.0
    # via black
psycopg2-binary==2.9.6
    # via -r requirements.in
pycodestyle==2.11.0
    # via flake8
pydantic==2.1.1
    # via fastapi
pydantic-core==2.4.0
    # via pydantic
pyflakes==3.1.0
    # via flake8
python-dotenv==1.0.0
    # via -r requirements.in
pytz==2023.3
    # via apscheduler
six==1.16.0
    # via apscheduler
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
sqlalchemy==2.0.19
    # via
    #   -r requirements.in
    #   alembic
starlette==0.27.0
    # via fastapi
tokenize-rt==5.1.0
    # via add-trailing-comma
typing-extensions==4.7.1
    # via
    #   alembic
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
tzlocal==5.0.1
    # via apscheduler
uvicorn==0.23.1
    # via -r requirements.in
yarl==1.9.2
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
